simplified voice settings UX
