+++ Packages/OsaurusCore/Resources/Localizable.xcstrings
@@ -1530,6 +1530,7 @@
 
     },
     "Affects pause detection, wake word detection, and voice activity thresholds" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -2320,6 +2321,12 @@
         }
       }
     },
+    "Applies to all voice modes: pause detection, wake word, and voice activity" : {
+
+    },
+    "Applies to both chat voice input and Transcription Mode" : {
+
+    },
     "Apply Theme" : {
       "localizations" : {
         "de" : {
@@ -2697,10 +2704,8 @@
     "Auto-discover relevant capabilities" : {
 
     },
-    "Auto-send disabled. You must manually send voice messages." : {
-
-    },
     "Auto-send is disabled in manual stop mode." : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -2717,6 +2722,7 @@
       }
     },
     "Auto-send or close voice input after this duration of silence" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -2788,6 +2794,15 @@
           }
         }
       }
+    },
+    "Auto-stop disabled. You must stop transcription manually." : {
+
+    },
+    "Auto-stop is disabled in manual stop mode." : {
+
+    },
+    "Auto-stop or close voice input after this duration of silence" : {
+
     },
     "Automate recurring AI tasks with custom schedules" : {
       "localizations" : {
@@ -4698,6 +4713,7 @@
       }
     },
     "Configure how voice messages are handled" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -11876,6 +11892,7 @@
       }
     },
     "Live Test" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -12757,9 +12774,6 @@
           }
         }
       }
-    },
-    "Message will prepare to send after %@ seconds of silence" : {
-
     },
     "Method" : {
       "localizations" : {
@@ -17756,6 +17770,7 @@
 
     },
     "RECORDING" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -21442,6 +21457,7 @@
       }
     },
     "Speak and see your words here..." : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -21858,6 +21874,9 @@
           }
         }
       }
+    },
+    "Stops after %@ seconds of silence" : {
+
     },
     "Stops anyone from pretending to be you." : {
 
@@ -22487,6 +22506,7 @@
       }
     },
     "Test your audio settings with live voice transcription" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -22798,6 +22818,7 @@
 
     },
     "These settings apply to all voice modes: Voice Input, Transcription, and VAD Mode." : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -23191,7 +23212,11 @@
         }
       }
     },
+    "Time to cancel before a chat message is automatically sent" : {
+
+    },
     "Time to cancel before message is automatically sent" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -23550,6 +23575,9 @@
           }
         }
       }
+    },
+    "Transcription Behavior" : {
+
     },
     "Transcription Mode" : {
       "localizations" : {
@@ -24805,6 +24833,7 @@
       }
     },
     "Voice Input" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -24853,6 +24882,7 @@
       }
     },
     "Voice Transcription" : {
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
