+++ Packages/OsaurusCore/Views/Voice/VoiceSetupTab.swift
@@ -14,6 +14,8 @@ struct VoiceSetupTab: View {
     @Environment(\.theme) private var theme
     @ObservedObject private var speechService = SpeechService.shared
     @ObservedObject private var modelManager = SpeechModelManager.shared
+    @ObservedObject private var audioInputManager = AudioInputManager.shared
+    @ObservedObject private var systemAudioManager = SystemAudioCaptureManager.shared
 
     /// Called when setup is complete
     var onComplete: (() -> Void)?
@@ -24,6 +26,21 @@ struct VoiceSetupTab: View {
     @State private var hasAppeared = false
     @State private var micButtonScale: CGFloat = 1.0
     @State private var isPressed = false
+
+    // Shared audio settings (apply to all voice modes)
+    @State private var sensitivity: VoiceSensitivity = .medium
+    @State private var hasLoadedSettings = false
+
+    private func loadSettings() {
+        sensitivity = SpeechConfigurationStore.load().sensitivity
+    }
+
+    private func saveSettings() {
+        var config = SpeechConfigurationStore.load()
+        config.sensitivity = sensitivity
+        SpeechConfigurationStore.save(config)
+        NotificationCenter.default.post(name: .voiceConfigurationChanged, object: nil)
+    }
 
     /// Whether all requirements are met
     private var isSetupComplete: Bool {
@@ -53,41 +70,63 @@ struct VoiceSetupTab: View {
     var body: some View {
         ScrollView {
             VStack(spacing: 0) {
+                // Onboarding hero — kept in a narrow, centered column
+                VStack(spacing: 0) {
+                    Spacer()
+                        .frame(height: 40)
+
+                    // Requirements checklist (compact)
+                    requirementsSection
+                        .opacity(hasAppeared ? 1 : 0)
+                        .offset(y: hasAppeared ? 0 : 8)
+                        .animation(.spring(response: 0.5, dampingFraction: 0.85).delay(0.05), value: hasAppeared)
+
+                    Spacer()
+                        .frame(height: 48)
+
+                    // Central voice test area
+                    voiceTestSection
+                        .opacity(hasAppeared ? 1 : 0)
+                        .scaleEffect(hasAppeared ? 1 : 0.95)
+                        .animation(.spring(response: 0.6, dampingFraction: 0.8).delay(0.15), value: hasAppeared)
+
+                    Spacer()
+                        .frame(height: 32)
+
+                    // Privacy footer
+                    privacyFooter
+                        .opacity(hasAppeared ? 1 : 0)
+                        .animation(.easeOut(duration: 0.4).delay(0.3), value: hasAppeared)
+                }
+                .padding(.horizontal, 24)
+                .frame(maxWidth: 520)
+                .frame(maxWidth: .infinity)
+
                 Spacer()
                     .frame(height: 40)
 
-                // Requirements checklist (compact)
-                requirementsSection
+                // Shared audio settings (sensitivity + input device) — full width,
+                // matching the other voice settings tabs.
+                audioSettingsSection
                     .opacity(hasAppeared ? 1 : 0)
-                    .offset(y: hasAppeared ? 0 : 8)
-                    .animation(.spring(response: 0.5, dampingFraction: 0.85).delay(0.05), value: hasAppeared)
-
-                Spacer()
-                    .frame(height: 48)
-
-                // Central voice test area
-                voiceTestSection
-                    .opacity(hasAppeared ? 1 : 0)
-                    .scaleEffect(hasAppeared ? 1 : 0.95)
-                    .animation(.spring(response: 0.6, dampingFraction: 0.8).delay(0.15), value: hasAppeared)
-
-                Spacer()
-                    .frame(height: 32)
-
-                // Privacy footer
-                privacyFooter
-                    .opacity(hasAppeared ? 1 : 0)
-                    .animation(.easeOut(duration: 0.4).delay(0.3), value: hasAppeared)
+                    .animation(.easeOut(duration: 0.4).delay(0.35), value: hasAppeared)
+                    .padding(.horizontal, 24)
 
                 Spacer()
             }
-            .padding(.horizontal, 24)
-            .frame(maxWidth: 520)
-            .frame(maxWidth: .infinity)
         }
         .onAppear {
+            if !hasLoadedSettings {
+                loadSettings()
+                hasLoadedSettings = true
+            }
             withAnimation {
                 hasAppeared = true
+            }
+        }
+        .onReceive(NotificationCenter.default.publisher(for: .voiceConfigurationChanged)) { _ in
+            if !isTestingVoice {
+                loadSettings()
             }
         }
         .onChange(of: speechService.currentTranscription) { _, newValue in
@@ -378,6 +417,287 @@ struct VoiceSetupTab: View {
                 .font(.system(size: 11))
                 .foregroundColor(theme.tertiaryText)
         }
+    }
+
+    // MARK: - Audio Settings Section
+
+    private var audioSettingsSection: some View {
+        VStack(spacing: 16) {
+            inputDeviceCard
+            sensitivitySettingsCard
+        }
+    }
+
+    private var sensitivitySettingsCard: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack(spacing: 12) {
+                ZStack {
+                    RoundedRectangle(cornerRadius: 12)
+                        .fill(theme.accentColor.opacity(0.15))
+                    Image(systemName: "waveform")
+                        .font(.system(size: 20, weight: .medium))
+                        .foregroundColor(theme.accentColor)
+                }
+                .frame(width: 48, height: 48)
+
+                VStack(alignment: .leading, spacing: 4) {
+                    Text("Voice Sensitivity", bundle: .module)
+                        .font(.system(size: 15, weight: .semibold))
+                        .foregroundColor(theme.primaryText)
+
+                    Text("Adjust how sensitive voice detection is", bundle: .module)
+                        .font(.system(size: 12))
+                        .foregroundColor(theme.secondaryText)
+                }
+
+                Spacer()
+            }
+
+            // Sensitivity Picker
+            VStack(alignment: .leading, spacing: 8) {
+                Text("Sensitivity Level", bundle: .module)
+                    .font(.system(size: 11, weight: .medium))
+                    .foregroundColor(theme.secondaryText)
+
+                HStack(spacing: 0) {
+                    ForEach(VoiceSensitivity.allCases, id: \.self) { level in
+                        Button(action: {
+                            sensitivity = level
+                            saveSettings()
+                        }) {
+                            Text(level.displayName)
+                                .font(.system(size: 13, weight: sensitivity == level ? .semibold : .medium))
+                                .foregroundColor(
+                                    sensitivity == level
+                                        ? (theme.isDark ? theme.primaryBackground : .white)
+                                        : theme.primaryText
+                                )
+                                .frame(maxWidth: .infinity)
+                                .padding(.vertical, 8)
+                                .contentShape(Rectangle())
+                                .background(
+                                    RoundedRectangle(cornerRadius: 6)
+                                        .fill(sensitivity == level ? theme.accentColor : Color.clear)
+                                )
+                        }
+                        .buttonStyle(.plain)
+                        .contentShape(Rectangle())
+                    }
+                }
+                .padding(4)
+                .background(
+                    RoundedRectangle(cornerRadius: 10)
+                        .fill(theme.tertiaryBackground)
+                )
+
+                Text(sensitivity.description)
+                    .font(.system(size: 11))
+                    .foregroundColor(theme.tertiaryText)
+            }
+
+            // Additional info
+            HStack(spacing: 8) {
+                Image(systemName: "info.circle")
+                    .font(.system(size: 12))
+                    .foregroundColor(theme.accentColor)
+                Text("Applies to all voice modes: pause detection, wake word, and voice activity", bundle: .module)
+                    .font(.system(size: 11))
+                    .foregroundColor(theme.secondaryText)
+            }
+            .padding(12)
+            .background(
+                RoundedRectangle(cornerRadius: 8)
+                    .fill(theme.accentColor.opacity(0.1))
+            )
+        }
+        .padding(20)
+        .background(
+            RoundedRectangle(cornerRadius: 16)
+                .fill(theme.cardBackground)
+                .overlay(
+                    RoundedRectangle(cornerRadius: 16)
+                        .stroke(theme.cardBorder, lineWidth: 1)
+                )
+        )
+    }
+
+    private var inputDeviceCard: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack(spacing: 12) {
+                ZStack {
+                    RoundedRectangle(cornerRadius: 12)
+                        .fill(theme.accentColor.opacity(0.15))
+                    Image(systemName: audioInputManager.selectedInputSource.iconName)
+                        .font(.system(size: 20, weight: .medium))
+                        .foregroundColor(theme.accentColor)
+                }
+                .frame(width: 48, height: 48)
+
+                VStack(alignment: .leading, spacing: 4) {
+                    Text("Audio Input", bundle: .module)
+                        .font(.system(size: 15, weight: .semibold))
+                        .foregroundColor(theme.primaryText)
+
+                    Text(inputSourceDescription)
+                        .font(.system(size: 12))
+                        .foregroundColor(theme.secondaryText)
+                }
+
+                Spacer()
+
+                Button(action: { audioInputManager.refreshDevices() }) {
+                    Image(systemName: "arrow.clockwise")
+                        .font(.system(size: 14, weight: .medium))
+                        .foregroundColor(theme.secondaryText)
+                        .padding(8)
+                        .background(
+                            RoundedRectangle(cornerRadius: 8)
+                                .fill(theme.tertiaryBackground)
+                        )
+                }
+                .buttonStyle(PlainButtonStyle())
+                .localizedHelp("Refresh available devices")
+            }
+
+            // Input Source Picker
+            VStack(alignment: .leading, spacing: 8) {
+                Text("Input Source", bundle: .module)
+                    .font(.system(size: 11, weight: .medium))
+                    .foregroundColor(theme.secondaryText)
+
+                HStack(spacing: 8) {
+                    ForEach(AudioInputSource.allCases, id: \.self) { source in
+                        let isSelected = audioInputManager.selectedInputSource == source
+                        let isDisabled = source == .systemAudio && !systemAudioManager.isAvailable
+
+                        Button(action: {
+                            if !isDisabled {
+                                withAnimation(.easeInOut(duration: 0.2)) {
+                                    audioInputManager.selectedInputSource = source
+                                }
+                            }
+                        }) {
+                            HStack(spacing: 6) {
+                                Image(systemName: source.iconName)
+                                    .font(.system(size: 12, weight: .medium))
+                                Text(source.displayName)
+                                    .font(.system(size: 12, weight: .medium))
+                            }
+                            .foregroundColor(
+                                isDisabled
+                                    ? theme.tertiaryText
+                                    : (isSelected
+                                        ? (theme.isDark ? theme.primaryBackground : .white) : theme.primaryText)
+                            )
+                            .padding(.horizontal, 14)
+                            .padding(.vertical, 8)
+                            .background(
+                                RoundedRectangle(cornerRadius: 8)
+                                    .fill(isSelected ? theme.accentColor : theme.tertiaryBackground)
+                            )
+                        }
+                        .buttonStyle(PlainButtonStyle())
+                        .disabled(isDisabled)
+                    }
+
+                    Spacer()
+                }
+            }
+
+            // Device picker (microphone mode only)
+            if audioInputManager.selectedInputSource == .microphone {
+                VStack(alignment: .leading, spacing: 8) {
+                    Text("Select Input Device", bundle: .module)
+                        .font(.system(size: 11, weight: .medium))
+                        .foregroundColor(theme.secondaryText)
+
+                    Menu {
+                        Button(action: { audioInputManager.selectDevice(nil) }) {
+                            HStack {
+                                Text("System Default", bundle: .module)
+                                if audioInputManager.selectedDeviceId == nil {
+                                    Image(systemName: "checkmark")
+                                }
+                            }
+                        }
+
+                        Divider()
+
+                        ForEach(audioInputManager.availableDevices) { device in
+                            Button(action: { audioInputManager.selectDevice(device.id) }) {
+                                HStack {
+                                    Text(device.name)
+                                    if device.isDefault {
+                                        Text("(Default)", bundle: .module)
+                                            .foregroundColor(.secondary)
+                                    }
+                                    if audioInputManager.selectedDeviceId == device.id {
+                                        Image(systemName: "checkmark")
+                                    }
+                                }
+                            }
+                        }
+                    } label: {
+                        HStack(spacing: 10) {
+                            Image(systemName: "mic.fill")
+                                .font(.system(size: 14))
+                                .foregroundColor(theme.accentColor)
+
+                            Text(selectedDeviceName)
+                                .font(.system(size: 13, weight: .medium))
+                                .foregroundColor(theme.primaryText)
+                                .lineLimit(1)
+
+                            Spacer()
+
+                            Image(systemName: "chevron.up.chevron.down")
+                                .font(.system(size: 10, weight: .medium))
+                                .foregroundColor(theme.tertiaryText)
+                        }
+                        .padding(.horizontal, 12)
+                        .padding(.vertical, 10)
+                        .background(
+                            RoundedRectangle(cornerRadius: 10)
+                                .fill(theme.inputBackground)
+                                .overlay(
+                                    RoundedRectangle(cornerRadius: 10)
+                                        .stroke(theme.inputBorder, lineWidth: 1)
+                                )
+                        )
+                    }
+                    .menuStyle(.borderlessButton)
+                }
+            }
+        }
+        .padding(20)
+        .background(
+            RoundedRectangle(cornerRadius: 16)
+                .fill(theme.cardBackground)
+                .overlay(
+                    RoundedRectangle(cornerRadius: 16)
+                        .stroke(theme.cardBorder, lineWidth: 1)
+                )
+        )
+    }
+
+    private var inputSourceDescription: String {
+        switch audioInputManager.selectedInputSource {
+        case .microphone:
+            return audioInputManager.selectedDevice?.name ?? L("System Default")
+        case .systemAudio:
+            return systemAudioManager.hasPermission
+                ? L("Computer audio")
+                : L("Permission required")
+        }
+    }
+
+    private var selectedDeviceName: String {
+        if let selectedId = audioInputManager.selectedDeviceId,
+            let device = audioInputManager.availableDevices.first(where: { $0.id == selectedId })
+        {
+            return device.name
+        }
+        return L("System Default")
     }
 
     // MARK: - Actions
